Migrate the functions in Tracker to DeviceLink
There's no benefit to the extra module now that it's asking the
DeviceLink if the device is online.
